현금 결제 부분에 myHSP->IgawAdbrixBuy 추가, 리와인드 속도 소수점 단위로 할 수 있도록

+++ Classes/StarGoldData.h
@@ -1013,7 +1013,7 @@ public:
 	void setIosHideVer(string t_str);
 	string getIosHideVer();
 	
-	int rewind_cnt_per_frame;
+	double rewind_cnt_per_frame;
 	
 private:
